Refactor system tests into a separate file
